split stream card to ethereum card and wrapper
